mhu_ped_oauth_provider/users/api/viewsets.py
from django.conf import settings
from django.contrib.auth.models import Group
from oauth2_provider.contrib.rest_framework import TokenHasReadWriteScope, TokenHasScope
from rest_framework import viewsets, permissions

from mhu_ped_oauth_provider.users.api.serializers import UserSerializer, GroupSerializer

User = settings.AUTH_USER_MODEL


class UserViewSet(viewsets.ModelViewSet):
    permission_classes = [permissions.IsAuthenticated, TokenHasReadWriteScope]
    serializer_class = UserSerializer

    def get_queryset(self):
        return User.objects.all()

class GroupViewSet(viewsets.ModelViewSet):
    permission_classes = [permissions.IsAuthenticated, TokenHasScope]
    required_scopes = ['groups']
    queryset = Group.objects.all()
    serializer_class = GroupSerializer
