requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/requirements.txt requirements/requirements.in
#
appnope==0.1.0            # via ipython
argon2-cffi==18.1.0       # via django
backcall==0.1.0           # via ipython
certifi==2018.4.16        # via requests
cffi==1.11.5              # via argon2-cffi
chardet==3.0.4            # via requests
collectfast==0.6.2
decorator==4.3.0          # via ipython, traitlets
defusedxml==0.5.0         # via python3-openid
django-allauth==0.36.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==2.0.7
django-model-utils==3.1.2
django-oauth-toolkit==1.1.2
django-storages==1.6.6    # via collectfast
django[argon2]==2.0.5
djangorestframework==3.8.2
gunicorn==19.8.1
idna==2.6                 # via requests
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==6.4.0            # via ipdb
jedi==0.12.0              # via ipython
oauthlib==2.1.0           # via django-oauth-toolkit, requests-oauthlib
parso==0.2.1              # via jedi
pexpect==4.5.0            # via ipython
pickleshare==0.7.4        # via ipython
prompt-toolkit==1.0.15    # via ipython
psycopg2-binary==2.7.4
ptyprocess==0.5.2         # via pexpect
pycparser==2.18           # via cffi
pygments==2.2.0           # via ipython
python3-openid==3.1.0     # via django-allauth
pytz==2018.4
raven==6.8.0
requests-oauthlib==0.8.0  # via django-allauth
requests==2.18.4          # via django-allauth, django-oauth-toolkit, requests-oauthlib
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via argon2-cffi, django-environ, django-extensions, prompt-toolkit, traitlets
sqlparse==0.2.4           # via django-debug-toolbar
traitlets==4.3.2          # via ipython
urllib3==1.22             # via requests
wcwidth==0.1.7            # via prompt-toolkit
